moved BitFieldExtract::generate function to header file
