+++ shared/rocroller/lib/include/rocRoller/CodeGen/Arithmetic/BitFieldExtract.hpp
@@ -76,3 +76,5 @@ namespace rocRoller
             = concatenate("BitFieldExtractGenerator<", toString(DATATYPE), ">");
     };
 }
+
+#include <rocRoller/CodeGen/Arithmetic/BitFieldExtract_impl.hpp>
